Breakout header file from sensors.cpp and delete unnecessary #includes.
